ai/requirements.txt
# AI/ML Dependencies
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.7.2
seaborn==0.12.2

# FastAPI for AI service
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Database connections
pymysql==1.1.0
elasticsearch==8.11.0

# Image processing and food recognition
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2

# Data processing
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
jupyter==1.0.0
notebook==7.0.6
